requirements.txt
streamlit
ultralytics
deep_sort_realtime
opencv-python-headless
Pillow
torch